For the Sections Labeled "Section Title - 1" I began changing the variable names from "Variable.Name" to "Variable.Name1"
In Section "Lamella Design-1 "
I finished the code to reassign N.SedTanks and N.Sedbays in the event of sed tank failure. 

In Section "Sedimentation Influent Manifold Program"
I changed the assignment of L.SedEst to reflect the case that there is only one sed tank with one bay and L.SedEst is still less than L.SedMin.  In this case we default to one sed tank with one bay of length L.SedMin.

git-svn-id: https://forge.cornell.edu/svn/repos/automated_design@3528 d22a8b0d-b447-0410-a14f-ca4c0a428a39